- Refactor: 'pages' -> 'features', added app.config.ts for standalone components, update routing - Added eslint
