+++ front/src/app/app.config.ts
@@ -1,0 +1,12 @@
+import { ApplicationConfig } from '@angular/core';
+import { provideRouter } from '@angular/router';
+import { provideHttpClient } from '@angular/common/http';
+import { routes } from './app.routes';
+
+export const appConfig: ApplicationConfig = {
+  providers: [
+    provideRouter(routes),
+    provideHttpClient(),
+    // Ajoute ici d'autres providers si nécessaire (ex: `provideAnimations()`)
+  ],
+};
